chore: dynamic token list init
